# runtime_stack/condenser/docs/swagger.yaml
definitions:
  container.CreateContainerRequest:
    properties:
      command:
        example:
        - /bin/sh
        - -c
        - echo hello; sleep 60
        items:
          type: string
        type: array
      env:
        items:
          type: string
        type: array
      image:
        example: alpine:latest
        type: string
      mount:
        example:
        - /host/dir:/container/dir
        - /src:/dst
        items:
          type: string
        type: array
      name:
        example: my-container
        type: string
      network:
        example: raind0
        type: string
      podId:
        example: pod-1234
        type: string
      port:
        example:
        - 8080:80
        - 4443:443
        items:
          type: string
        type: array
      tty:
        example: false
        type: boolean
    type: object
  container.ExecContainerRequest:
    properties:
      command:
        example:
        - /bin/sh
        - -c
        - echo hello
        items:
          type: string
        type: array
      tty:
        example: true
        type: boolean
    type: object
  container.StartContainerRequest:
    properties:
      tty:
        example: false
        type: boolean
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        description: success | fail
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/containers:
    get:
      description: get all container list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get container list
      tags:
      - containers
    post:
      consumes:
      - application/json
      description: create a new container
      parameters:
      - description: Container Spec
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.CreateContainerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: Create a container
      tags:
      - containers
  /v1/containers/{containerId}:
    get:
      description: get an exitsting container info
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get container info
      tags:
      - containers
  /v1/containers/{containerId}/actions/delete:
    delete:
      description: delete an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: delete a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/exec:
    post:
      description: execute command inside an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      - description: Execute Options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.ExecContainerRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: exec a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/start:
    post:
      description: start an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      - description: Start Options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/container.StartContainerRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: start a container
      tags:
      - containers
  /v1/containers/{containerId}/actions/stop:
    post:
      description: stop an exitsting container
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: stop a container
      tags:
      - containers
  /v1/containers/{containerId}/log:
    get:
      description: get container log
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get container log
      tags:
      - containers
  /v1/containers/{containerId}/logpath:
    get:
      description: get container log path
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get container log path
      tags:
      - containers
  /v1/containers/{containerId}/stats:
    get:
      description: get container stats
      parameters:
      - description: Container ID
        in: path
        name: containerId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      summary: get container stats
      tags:
      - containers
...
swagger: "2.0"